--- src/entities/index.ts
@@ -1,2 +1,0 @@
-export { PostHeading, getPostApi, type Post } from './post';
-export { Card, type CardType } from './card';
