--- src/features/index.ts
@@ -1,8 +1,0 @@
-export { GetRandomPost } from './get-random-post';
-export { AlertInputText } from './alert-input-text';
-export {
-  NavigationWithPermissions,
-  generateNavigationListWithPermissions,
-  navigationList,
-  checkHasUserPermission,
-} from './navigation-with-permissions';
