--- src/features/navigation-with-permissions/index.ts
@@ -1,4 +1,0 @@
-export { NavigationWithPermissions } from './ui/navigation-with-permissions';
-export { generateNavigationListWithPermissions } from './helpers/generate-navigation-list-with-permissions';
-export { checkHasUserPermission } from './helpers/check-has-user-permission';
-export { navigationList } from './const/navigation-list';
