--- src/features/navigation-with-permissions/ui/navigation-with-permissions.module.scss
@@ -1,33 +1,0 @@
-.navigation {
-  padding: 10px;
-  border: 1px solid #ccc;
-}
-
-.level_1 {
-  margin: 20px 0;
-  color: red;
-}
-
-.level_2 {
-  padding-left: 20px;
-  margin: 10px 0;
-  color: gold;
-}
-
-.level_3 {
-  padding-left: 20px;
-  margin: 10px 0;
-  color: blue;
-}
-
-.level_4 {
-  padding-left: 20px;
-  margin: 10px 0;
-  color: green;
-}
-
-.level_5 {
-  padding-left: 20px;
-  margin: 10px 0;
-  color: orange;
-}
