+++ src/features/navigation/const/types.ts
@@ -1,0 +1,12 @@
+export type Route = {
+  name: string;
+  pathname: string;
+  getLink: () => string;
+  text: string;
+};
+
+export type NavigationItem = {
+  name: string;
+  text: string;
+  children: Array<Route | NavigationItem>;
+};
