+++ src/features/navigation/index.ts
@@ -1,0 +1,2 @@
+export { Navigation } from './ui/navigation';
+export { type Route, type NavigationItem } from './const/types';
