+++ src/features/navigation/ui/navigation.module.scss
@@ -1,0 +1,9 @@
+.navigation {
+  padding: 10px;
+  border: 1px solid #ccc;
+}
+
+.navigation_item {
+  padding-left: 20px;
+  margin: 10px 0;
+}
