+++ src/pages/navigation/const/user-read-permissions.ts
