+++ src/pages/navigation/helpers/check-has-user-permission.ts
