+++ src/pages/random-post/ui/random-post-page.tsx
@@ -1,4 +1,4 @@
-import { PostCard } from '@widgets';
+import { PostCard } from '@widgets/post-card';
 import styles from './random-post-page.module.scss';
 
 export const RandomPostPage = () => (
