--- src/widgets/index.ts
@@ -1,4 +1,0 @@
-export { Header } from './header';
-export { PostCard } from './post-card';
-export { InterestingFacts } from './interesting-facts';
-export { CardsList } from './cards-list';
