+++ tsconfig.json
@@ -8,11 +8,11 @@
     "allowSyntheticDefaultImports": true,
     "baseUrl": "./src",
     "paths": {
-      "@entities": ["entities/"],
-      "@features": ["features/"],
+      "@entities/*": ["entities/*"],
+      "@features/*": ["features/*"],
       "@pages": ["pages/"],
       "@shared": ["shared/"],
-      "@widgets": ["widgets/"]
+      "@widgets/*": ["widgets/*"]
     }
   },
   "include": ["src"]
